+++ OgreMain/src/Animation/OgreSkeletonInstance.cpp
@@ -296,11 +296,11 @@ namespace Ogre
         const BoneTransform &firstBoneTransf = firstBone._getTransform();
 
         // Get bone offset relative to the beginning of its depth level.
-        uintptr_t diff = ( boneTransf.mOwner - firstBoneTransf.mOwner ) + boneTransf.mIndex;
+        ptrdiff_t diff = ( boneTransf.mOwner - firstBoneTransf.mOwner ) + boneTransf.mIndex;
         // Offset by all past bones from parent levels.
         diff += mDefinition->getNumberOfBoneBlocks( depthLevel ) * ARRAY_PACKED_REALS;
 
-        assert( diff < mManualBones.size() * ARRAY_PACKED_REALS &&
+        assert( static_cast<size_t>( diff ) < mManualBones.size() * ARRAY_PACKED_REALS &&
                 "Offset incorrectly calculated. manualBones[diff] will overflow!" );
 
         Real *manualBones = reinterpret_cast<Real *>( mManualBones.get() );
@@ -318,11 +318,11 @@ namespace Ogre
         const BoneTransform &firstBoneTransf = firstBone._getTransform();
 
         // Get bone offset relative to the beginning of its depth level.
-        uintptr_t diff = ( boneTransf.mOwner - firstBoneTransf.mOwner ) + boneTransf.mIndex;
+        ptrdiff_t diff = ( boneTransf.mOwner - firstBoneTransf.mOwner ) + boneTransf.mIndex;
         // Offset by all past bones from parent levels.
         diff += mDefinition->getNumberOfBoneBlocks( depthLevel ) * ARRAY_PACKED_REALS;
 
-        assert( diff < mManualBones.size() * ARRAY_PACKED_REALS &&
+        assert( static_cast<size_t>( diff ) < mManualBones.size() * ARRAY_PACKED_REALS &&
                 "Offset incorrectly calculated. manualBones[diff] will overflow!" );
 
         const Real *manualBones = reinterpret_cast<const Real *>( mManualBones.get() );
